Add pipeline
We didn't separate the tests from the deploy because we don't intend to
be interacting with this repo often. We would like to migrate away from
it.

We also created a user instead of a client because the code only
supports user-based auth and we are trying to keep our changeset as
minimal as possible.

[#166781016]

Signed-off-by: Aaron Hurley <ahurley@pivotal.io>