Use --guid flag when checking for app presence
A normal `cf app` invocation makes numerous API calls in order to get
all of the information it needs to populate the app. This behavior
meant that the canary apps that had failed to stage after they had had
the stack-auditor plugin run on them caused `cf app` to fail because it
couldn't get their app instance information.

In turn, the failed command meant that `canaries` would consider the app
as "not existing" and would try and push it fresh, but the apps' actual
configuration had diverged from what the configuration that was
hard-coded inside the gem.

[#166781016]

Signed-off-by: Tom Viehman <tviehman@pivotal.io>